+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-
-### Updated
+## [4.11.1] - 2020-05-4
+### Fixed
 - Upload viewport image to Azure cloud for visual locators. [Trello 1672](https://trello.com/c/9r6TeoVo)
 ### Added
 - Disable SSL verification. Accept all certificates. [Trello 1771](https://trello.com/c/WPWFJFcD)